stdint

+++ include/cmds.h
@@ -1,8 +1,8 @@
 #ifndef INCLUDE_CMDS_H
 #define INCLUDE_CMDS_H
 
-int init(int argc);
-int load(int argc);
+int init(uint16_t argc);
+int load(uint16_t argc);
 void help(void);
 
 //#pragma message __FILE__
